pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "footpredict"
version = "0.1.0"
description = "Football match prediction + Kelly staking + dashboard"
requires-python = ">=3.11"
dependencies = [
  "requests>=2.31.0",
  "pandas>=2.2.0",
  "numpy>=1.26.0",
  "pyarrow>=16.0.0",
  "scikit-learn>=1.4.0",
  "joblib>=1.3.0",
  "python-dotenv>=1.0.0",
  "streamlit>=1.35.0"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["api*", "data*", "features*", "models*", "betting*", "validation*", "utils*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]